# api/swagger.json
{"openapi":"3.0.0","paths":{"/unidade":{"get":{"operationId":"UnidadeController_findAll","parameters":[{"name":"unidade_tipo_id","required":false,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Lista Todas as Unidades","tags":["Unidade"]}},"/unidade/tipo":{"get":{"operationId":"UnidadeTipoController_findAll","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Lista todos os tipos de unidade","tags":["UnidadeTipo"]}},"/causa":{"get":{"operationId":"CausaController_findAll","parameters":[{"name":"tipo","required":false,"in":"query","schema":{"type":"string","enum":["EXTERNA","QUEIXA"]}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Lista todas as causas","tags":["Causa"]}},"/causa/sg":{"get":{"operationId":"CausaSgController_findAll","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Lista todas as causas consideradas srags cadastradas","tags":["CausaSg"]}},"/cid/grupo":{"get":{"operationId":"CidGrupoController_findAll","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Lista todos os grupos de cid cadastrados","tags":["CidGrupo"]}},"/cid/categoria":{"get":{"operationId":"CidCategoriaController_findAll","parameters":[{"name":"cid_grupo_id","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Lista todas as categorias de cid cadastradas","tags":["CidCategoria"]}},"/cid":{"get":{"operationId":"CidController_findAll","parameters":[{"name":"cid_categoria_id","required":false,"in":"query","schema":{"type":"number"}},{"name":"codigo","required":false,"in":"query","schema":{"type":"string"}},{"name":"take","required":false,"in":"query","schema":{"type":"number"}},{"name":"skip","required":false,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Lista todas as CIDs cadastradas por categoria","tags":["Cid"]}},"/cid/srag":{"get":{"operationId":"CidSragController_findAll","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Lista todos os grupos de cid cadastrados","tags":["CidSrag"]}},"/usuario/{id}":{"get":{"operationId":"UsuarioController_getById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Busca um usuario pelo id no banco","tags":["Usuario"]}},"/auth/signup":{"post":{"operationId":"AuthController_signUp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthSignUpDto"}}}},"responses":{"201":{"description":""}},"tags":["Autenticação"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthLoginDto"}}}},"responses":{"201":{"description":""}},"tags":["Autenticação"]}},"/auth/logout":{"post":{"operationId":"AuthController_logout","parameters":[],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["Autenticação"]}},"/boletim/internacao-admissao":{"post":{"operationId":"BoletimInternacaoAdmissaoController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BoletimInternacaoAdmissaoCreateDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["BoletimInternacaoAdmissao"]}},"/boletim/internacao-ocupacao":{"post":{"operationId":"BoletimInternacaoOcupacaoController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BoletimInternacaoOcupacaoCreateDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["BoletimInternacaoOcupacao"]}},"/boletim/sindrome-gripal":{"get":{"operationId":"BoletimSindromeGripalController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BoletimSindromeGripalFindAllResponse"}}}}}},"security":[{"bearer":[]}],"tags":["BoletimSindromeGripal"]},"post":{"operationId":"BoletimSindromeGripalController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BoletimSindromeGripalCreateDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["BoletimSindromeGripal"]}},"/boletim/sindrome-gripal/{id}":{"get":{"operationId":"BoletimSindromeGripalController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BoletimSindromeGripalFindOneResponse"}}}}},"security":[{"bearer":[]}],"tags":["BoletimSindromeGripal"]}},"/ficha-spa":{"post":{"operationId":"FichaSpaController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FichaSpaCreateDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FichaSpaFindOneResponse"}}}}},"security":[{"bearer":[]}],"tags":["FichaSpa"]},"get":{"operationId":"FichaSpaController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FichaSpaFindAllResponse"}}}}}},"security":[{"bearer":[]}],"tags":["FichaSpa"]}},"/ficha-spa/{id}":{"get":{"operationId":"FichaSpaController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FichaSpaFindOneResponse"}}}}},"security":[{"bearer":[]}],"tags":["FichaSpa"]}}},"info":{"title":"e-SEMUS","description":"O e-SEMUS (Sistema Integrado de SEMUS) é um sistema de informação voltado para a gestão e integração de dados da Secretaria Municipal de Saude (SEMUS)","version":"1.0","contact":{}},"tags":[{"name":"Autenticação","description":""}],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http","name":"authorization","description":"Token para autenticação do usuario","in":"header"}},"schemas":{"AuthSignUpDto":{"type":"object","properties":{"setor_area_cargo_id":{"type":"number"},"nome":{"type":"string"},"senha":{"type":"string"},"matricula":{"type":"string"},"email":{"type":"string","format":"email"},"cpf":{"type":"string"}},"required":["setor_area_cargo_id","nome","senha","matricula","email","cpf"]},"AuthLoginDto":{"type":"object","properties":{"usuario":{"type":"string"},"senha":{"type":"string"}},"required":["usuario","senha"]},"BoletimInternacaoAdmissaoItemData":{"type":"object","properties":{"boletim_internacao_admissao_id":{"type":"number"},"numero_de_admissoes":{"type":"number"},"cid_id":{"type":"number"},"faixa_etaria":{"type":"string"},"genero":{"type":"object"}},"required":["boletim_internacao_admissao_id","numero_de_admissoes","cid_id","faixa_etaria","genero"]},"BoletimInternacaoAdmissaoCreateDto":{"type":"object","properties":{"unidade_id":{"type":"number"},"referente_ao_dia":{"format":"date-time","type":"string"},"items":{"type":"array","items":{"$ref":"#/components/schemas/BoletimInternacaoAdmissaoItemData"}}},"required":["unidade_id","referente_ao_dia","items"]},"BoletimInternacaoOcupacaoCreateDto":{"type":"object","properties":{"unidade_id":{"type":"number"},"referente_ao_dia":{"format":"date-time","type":"string"},"pacientes_uti":{"type":"number"},"pacientes_uti_isolamento":{"type":"number"},"pacientes_enfermaria":{"type":"number"},"pacientes_enfermaria_isolamento":{"type":"number"},"pacientes_lsvp":{"type":"number"},"pacientes_estabilizacao_vermelha":{"type":"number"}},"required":["unidade_id","referente_ao_dia","pacientes_uti","pacientes_uti_isolamento","pacientes_enfermaria","pacientes_enfermaria_isolamento","pacientes_lsvp","pacientes_estabilizacao_vermelha"]},"BoletimSindromeGripalFindAllResponse":{"type":"object","properties":{"referente_ao_dia":{"format":"date-time","type":"string"},"id":{"type":"number"},"unidade":{"type":"object","properties":{"id":{"type":"number"},"nome":{"type":"string"}},"required":["id","nome"]},"usuario":{"type":"object","properties":{"id":{"type":"string"},"criado_em":{"format":"date-time","type":"string","nullable":true},"atualizado_em":{"format":"date-time","type":"string","nullable":true},"nome":{"type":"string"},"matricula":{"type":"string"},"unidade_lotada_id":{"type":"number"},"setor_area_cargo_id":{"type":"number"},"cpf":{"type":"string"},"email":{"type":"string","nullable":true}},"required":["id","criado_em","atualizado_em","nome","matricula","unidade_lotada_id","setor_area_cargo_id","cpf","email"]},"criado_em":{"format":"date-time","type":"string","nullable":true},"atualizado_em":{"format":"date-time","type":"string","nullable":true}},"required":["referente_ao_dia","id","unidade","usuario","criado_em","atualizado_em"]},"BoletimSindromeGripalCreateDto":{"type":"object","properties":{"referente_ao_dia":{"format":"date-time","type":"string"},"total_atendimentos_sd_com_queixa_gripal":{"type":"number"},"total_atendimentos_sd_sem_queixa_gripal":{"type":"number"},"total_atendimentos_sn_com_queixa_gripal":{"type":"number"},"total_atendimentos_sn_sem_queixa_gripal":{"type":"number"},"total_internacoes_apos_atendimento_urgencia_com_queixa_gripal":{"type":"number"},"total_internacoes_apos_atendimento_urgencia_sem_queixa_gripal":{"type":"number"},"total_obitos":{"type":"number"},"total_transferencias":{"type":"number"}},"required":["referente_ao_dia","total_atendimentos_sd_com_queixa_gripal","total_atendimentos_sd_sem_queixa_gripal","total_atendimentos_sn_com_queixa_gripal","total_atendimentos_sn_sem_queixa_gripal","total_internacoes_apos_atendimento_urgencia_com_queixa_gripal","total_internacoes_apos_atendimento_urgencia_sem_queixa_gripal","total_obitos","total_transferencias"]},"BoletimSindromeGripalFindOneResponse":{"type":"object","properties":{"id":{"type":"number"},"unidade":{"type":"object","properties":{"nome":{"type":"string"},"sigla":{"type":"string","nullable":true}},"required":["nome","sigla"]},"usuario":{"type":"object","properties":{"nome":{"type":"string"},"setor_area_cargo":{"type":"object","properties":{"setor":{"type":"object","properties":{"nome":{"type":"string"}},"selfRequired":true},"area":{"type":"object","properties":{"nome":{"type":"string"}},"selfRequired":true},"cargo":{"type":"object","properties":{"nome":{"type":"string"}},"selfRequired":true}},"selfRequired":true},"matricula":{"type":"string"}},"required":["nome","setor_area_cargo","matricula"]},"unidade_id":{"type":"number"},"referente_ao_dia":{"format":"date-time","type":"string"},"usuario_responsavel_preenchimento_id":{"type":"string","nullable":true},"total_atendimentos_sd_com_queixa_gripal":{"type":"number"},"total_atendimentos_sd_sem_queixa_gripal":{"type":"number"},"total_atendimentos_sn_com_queixa_gripal":{"type":"number"},"total_atendimentos_sn_sem_queixa_gripal":{"type":"number"},"total_internacoes_apos_atendimento_urgencia_com_queixa_gripal":{"type":"number"},"total_internacoes_apos_atendimento_urgencia_sem_queixa_gripal":{"type":"number"},"total_obitos":{"type":"number"},"total_transferencias":{"type":"number"},"criado_em":{"format":"date-time","type":"string","nullable":true},"atualizado_em":{"format":"date-time","type":"string","nullable":true}},"required":["id","unidade","usuario","unidade_id","referente_ao_dia","usuario_responsavel_preenchimento_id","total_atendimentos_sd_com_queixa_gripal","total_atendimentos_sd_sem_queixa_gripal","total_atendimentos_sn_com_queixa_gripal","total_atendimentos_sn_sem_queixa_gripal","total_internacoes_apos_atendimento_urgencia_com_queixa_gripal","total_internacoes_apos_atendimento_urgencia_sem_queixa_gripal","total_obitos","total_transferencias","criado_em","atualizado_em"]},"FichaSpaRecepcao":{"type":"object","properties":{"hora_da_recepcao":{"format":"date-time","type":"string"},"nao_identificado":{"type":"boolean"},"nome_paciente":{"type":"string"},"genero":{"type":"object","default":"M"},"data_de_nascimento":{"format":"date-time","type":"string"},"cartao_sus_ou_cpf":{"type":"string"},"municipio_rg_id":{"type":"number"},"raca_cor":{"type":"object","default":"BRANCA"},"procedencia":{"type":"object","default":"DEMANDA_ESPONTANEA_SAO_LUIS"},"municipio_procedencia_interior_id":{"type":"number","nullable":true}},"required":["hora_da_recepcao","nao_identificado","nome_paciente","genero","data_de_nascimento","cartao_sus_ou_cpf","municipio_rg_id","raca_cor","procedencia","municipio_procedencia_interior_id"]},"FichaSpaDoencaPrexistente":{"type":"object","properties":{"has":{"type":"boolean"},"dm":{"type":"boolean"},"drc":{"type":"boolean"},"outros":{"type":"string","nullable":true}},"required":["has","dm","drc","outros"]},"FichaSpaProtocoloECondicaoEspecial":{"type":"object","properties":{"sepse":{"type":"boolean"},"dor_toracica":{"type":"boolean"},"avc":{"type":"boolean"},"notificacao":{"type":"boolean"},"notificacao_agravo_id":{"type":"number","nullable":true}},"required":["sepse","dor_toracica","avc","notificacao","notificacao_agravo_id"]},"FichaSpaEncaminhamento":{"type":"object","properties":{"encaminhado_para_1":{"type":"object","default":"ESTABILIZACAO"},"encaminhado_para_2":{"type":"object","default":"ECG"},"hora_da_realizacao_ecg":{"format":"date-time","type":"string","nullable":true}},"required":["encaminhado_para_1","encaminhado_para_2","hora_da_realizacao_ecg"]},"FichaSpaQueixa":{"type":"object","properties":{"causa_id":{"type":"number"}},"required":["causa_id"]},"FichaSpaClassificacao":{"type":"object","properties":{"hora_da_classificacao":{"format":"date-time","type":"string"},"causa_externa_id":{"type":"number"},"retornou_com_menos_ou_igual_48h":{"type":"boolean"},"doenca_prexistente":{"$ref":"#/components/schemas/FichaSpaDoencaPrexistente"},"protocolo_e_condicao_especial":{"$ref":"#/components/schemas/FichaSpaProtocoloECondicaoEspecial"},"encaminhamento":{"$ref":"#/components/schemas/FichaSpaEncaminhamento"},"manchester":{"type":"object","default":"VERMELHO"},"queixa":{"type":"array","items":{"$ref":"#/components/schemas/FichaSpaQueixa"}}},"required":["hora_da_classificacao","causa_externa_id","retornou_com_menos_ou_igual_48h","doenca_prexistente","protocolo_e_condicao_especial","encaminhamento","manchester","queixa"]},"FichaSpaExameSolicitado":{"type":"object","properties":{"raio_x":{"type":"boolean"},"tomografia":{"type":"boolean"},"ecg":{"type":"boolean"},"ultrassonografia":{"type":"boolean"},"laboratorio":{"type":"boolean"}},"required":["raio_x","tomografia","ecg","ultrassonografia","laboratorio"]},"FichaSpaAtendimentoMedico":{"type":"object","properties":{"horario_do_atendimento_medico":{"format":"date-time","type":"string"},"exame_solicitado":{"$ref":"#/components/schemas/FichaSpaExameSolicitado"}},"required":["horario_do_atendimento_medico","exame_solicitado"]},"FichaSpaPlanoTerapeutico":{"type":"object","properties":{"horario_do_atendimento_na_medicacao_observacao":{"format":"date-time","type":"string"}},"required":["horario_do_atendimento_na_medicacao_observacao"]},"FichaSpaDestinoFinalDoPaciente":{"type":"object","properties":{"horario_da_saida":{"format":"date-time","type":"string"},"destino_final":{"type":"object","default":"ALTA"},"unidade_transferida_id":{"type":"number"},"horario_da_saida_para_o_sertor_de_internacao":{"format":"date-time","type":"string"}},"required":["horario_da_saida","destino_final","unidade_transferida_id","horario_da_saida_para_o_sertor_de_internacao"]},"FichaSpaCreateDto":{"type":"object","properties":{"numero_da_ficha":{"type":"string"},"data_da_ficha":{"format":"date-time","type":"string"},"recepcao":{"$ref":"#/components/schemas/FichaSpaRecepcao"},"classificacao":{"$ref":"#/components/schemas/FichaSpaClassificacao"},"atendimento_medico":{"$ref":"#/components/schemas/FichaSpaAtendimentoMedico"},"plano_terapeutico":{"$ref":"#/components/schemas/FichaSpaPlanoTerapeutico"},"destino_final_do_paciente":{"$ref":"#/components/schemas/FichaSpaDestinoFinalDoPaciente"}},"required":["numero_da_ficha","data_da_ficha","recepcao","classificacao","atendimento_medico","plano_terapeutico","destino_final_do_paciente"]},"FichaSpaFindOneResponse":{"type":"object","properties":{"numero_da_ficha":{"type":"string"},"data_da_ficha":{"format":"date-time","type":"string"},"id":{"type":"number"},"usuario":{"type":"object","properties":{"nome":{"type":"string"}},"required":["nome"]},"unidade":{"type":"object","properties":{"id":{"type":"number"},"nome":{"type":"string"},"sigla":{"type":"string","nullable":true}},"required":["id","nome","sigla"]},"ficha_spa_recepcao":{"type":"array","items":{"type":"object","properties":{"hora_da_recepcao":{"format":"date-time","type":"string"},"nao_identificado":{"type":"boolean","nullable":true},"nome_paciente":{"type":"string","nullable":true},"genero":{"type":"object","nullable":true},"data_de_nascimento":{"format":"date-time","type":"string","nullable":true},"cartao_sus_ou_cpf":{"type":"string","nullable":true},"raca_cor":{"type":"object","nullable":true},"procedencia":{"type":"object","nullable":true},"criado_em":{"format":"date-time","type":"string","nullable":true},"atualizado_em":{"format":"date-time","type":"string","nullable":true},"municipio_procedencia_interior":{"type":"object","properties":{"id":{"type":"number"},"nome":{"type":"string"},"estado":{"type":"object","properties":{"id":{"type":"number"},"nome":{"type":"string"},"uf":{"type":"string"}},"selfRequired":true}},"selfRequired":true},"municipio_rg":{"type":"object","properties":{"id":{"type":"number"},"nome":{"type":"string"},"estado":{"type":"object","properties":{"id":{"type":"number"},"nome":{"type":"string"},"uf":{"type":"string"}},"selfRequired":true}},"selfRequired":true}},"required":["hora_da_recepcao","nao_identificado","nome_paciente","genero","data_de_nascimento","cartao_sus_ou_cpf","raca_cor","procedencia","criado_em","atualizado_em","municipio_procedencia_interior","municipio_rg"]}},"ficha_spa_classificacao":{"type":"array","items":{"type":"object","properties":{"hora_da_classificacao":{"format":"date-time","type":"string"},"retornou_com_menos_ou_igual_48h":{"type":"boolean"},"manchester":{"type":"object"},"causa":{"type":"object","properties":{"id":{"type":"number"},"criado_em":{"format":"date-time","type":"string","nullable":true},"atualizado_em":{"format":"date-time","type":"string","nullable":true},"nome":{"type":"string","nullable":true},"tipo":{"type":"object","nullable":true}},"selfRequired":true},"ficha_spa_doenca_preexistente":{"type":"array","items":{"type":"object","properties":{"has":{"type":"boolean"},"dm":{"type":"boolean"},"drc":{"type":"boolean"},"outros":{"type":"string","nullable":true},"id":{"type":"number"},"criado_em":{"format":"date-time","type":"string","nullable":true},"atualizado_em":{"format":"date-time","type":"string","nullable":true},"ficha_spa_classificacao_id":{"type":"number"}},"required":["has","dm","drc","outros","id","criado_em","atualizado_em","ficha_spa_classificacao_id"]},"selfRequired":true},"ficha_spa_encaminhamento":{"type":"array","items":{"type":"object","properties":{"encaminhado_para_1":{"type":"object"},"encaminhado_para_2":{"type":"object"},"hora_da_realizacao_ecg":{"format":"date-time","type":"string","nullable":true},"id":{"type":"number"},"criado_em":{"format":"date-time","type":"string","nullable":true},"atualizado_em":{"format":"date-time","type":"string","nullable":true},"ficha_spa_classificacao_id":{"type":"number","nullable":true}},"required":["encaminhado_para_1","encaminhado_para_2","hora_da_realizacao_ecg","id","criado_em","atualizado_em","ficha_spa_classificacao_id"]},"selfRequired":true},"ficha_spa_protocolo_e_condicao_especial":{"type":"array","items":{"type":"object","properties":{"sepse":{"type":"boolean"},"dor_toracica":{"type":"boolean"},"avc":{"type":"boolean"},"notificacao":{"type":"boolean"},"notificacao_agravo_id":{"type":"number","nullable":true},"id":{"type":"number"},"criado_em":{"format":"date-time","type":"string","nullable":true},"atualizado_em":{"format":"date-time","type":"string","nullable":true},"ficha_spa_classificacao_id":{"type":"number"}},"required":["sepse","dor_toracica","avc","notificacao","notificacao_agravo_id","id","criado_em","atualizado_em","ficha_spa_classificacao_id"]},"selfRequired":true}},"required":["hora_da_classificacao","retornou_com_menos_ou_igual_48h","manchester","causa"]}},"ficha_spa_atendimento_medico":{"type":"array","items":{"type":"object","properties":{"horario_do_atendimento_medico":{"format":"date-time","type":"string"},"id":{"type":"number"},"criado_em":{"format":"date-time","type":"string","nullable":true},"atualizado_em":{"format":"date-time","type":"string","nullable":true},"ficha_spa_id":{"type":"number"}},"required":["horario_do_atendimento_medico","id","criado_em","atualizado_em","ficha_spa_id"]}},"ficha_spa_plano_terapeutico":{"type":"array","items":{"type":"object","properties":{"horario_do_atendimento_na_medicacao_observacao":{"format":"date-time","type":"string","nullable":true},"id":{"type":"number"},"criado_em":{"format":"date-time","type":"string","nullable":true},"atualizado_em":{"format":"date-time","type":"string","nullable":true},"ficha_spa_id":{"type":"number"}},"required":["horario_do_atendimento_na_medicacao_observacao","id","criado_em","atualizado_em","ficha_spa_id"]}},"ficha_spa_destino_final_do_paciente":{"type":"array","items":{"type":"object","properties":{"horario_da_saida":{"format":"date-time","type":"string","nullable":true},"destino_final":{"type":"object","nullable":true},"unidade_transferida_id":{"type":"number","nullable":true},"id":{"type":"number"},"criado_em":{"format":"date-time","type":"string","nullable":true},"atualizado_em":{"format":"date-time","type":"string","nullable":true},"ficha_spa_id":{"type":"number"},"horario_da_saida_para_o_setor_de_internacao":{"format":"date-time","type":"string","nullable":true}},"required":["horario_da_saida","destino_final","unidade_transferida_id","id","criado_em","atualizado_em","ficha_spa_id","horario_da_saida_para_o_setor_de_internacao"]}},"criado_em":{"format":"date-time","type":"string","nullable":true},"atualizado_em":{"format":"date-time","type":"string","nullable":true}},"required":["numero_da_ficha","data_da_ficha","id","usuario","unidade","ficha_spa_recepcao","ficha_spa_classificacao","ficha_spa_atendimento_medico","ficha_spa_plano_terapeutico","ficha_spa_destino_final_do_paciente","criado_em","atualizado_em"]},"FichaSpaFindAllResponse":{"type":"object","properties":{"numero_da_ficha":{"type":"string"},"data_da_ficha":{"format":"date-time","type":"string"},"id":{"type":"number"},"usuario":{"type":"object","properties":{"id":{"type":"string"},"nome":{"type":"string"}},"required":["id","nome"]},"unidade":{"type":"object","properties":{"id":{"type":"number"},"nome":{"type":"string"},"sigla":{"type":"string","nullable":true}},"required":["id","nome","sigla"]},"criado_em":{"format":"date-time","type":"string","nullable":true},"atualizado_em":{"format":"date-time","type":"string","nullable":true}},"required":["numero_da_ficha","data_da_ficha","id","usuario","unidade","criado_em","atualizado_em"]}}}}